Move password hashing to PasswordHelper
